gateway: migrate annotation wrappers
